discord-bot/pyproject.toml
[project]
name = "ollama-discord-bot"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "discord.py>=2.3.2",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]
